+++ Sources/DangerFixtures/DangerDSLResources/DSLGitLabProjectMilestone.swift
@@ -1,0 +1,16 @@
+public let DSLGitLabProjectMilestoneJSON = """
+{
+   "id": 1672801,
+   "iid": 1,
+   "project_id": 21265694,
+   "title": "Project Milestone",
+   "description": "",
+   "state": "active",
+   "created_at": "2020-10-13T00:23:21.267Z",
+   "updated_at": "2020-10-13T00:23:21.267Z",
+   "due_date": null,
+   "start_date": null,
+   "expired": null,
+   "web_url": "https://gitlab.com/groupname/projectname/-/milestones/1"
+}
+"""