+++ Tests/DangerTests/BitbucketCloudResources/BitbucketCloudCommentUserWithoutAccountIdAndNickname.swift
@@ -1,0 +1,53 @@
+public let BitBucketCloudCommentUserWithoutAccountIdAndNickname = """
+{
+    "links": {
+      "self": {
+        "href": "https://api.bitbucket.org/2.0/repositories/Zagolo/danger-kotlin/pullrequests/1/comments/109197958"
+      },
+      "html": {
+        "href": "https://bitbucket.org/Zagolo/danger-kotlin/pull-requests/1/_/diff#comment-109197958"
+      }
+    },
+    "deleted": false,
+    "pullrequest": {
+      "type": "pullrequest",
+      "id": 1,
+      "links": {
+        "self": {
+          "href": "https://api.bitbucket.org/2.0/repositories/Zagolo/danger-kotlin/pullrequests/1"
+        },
+        "html": {
+          "href": "https://bitbucket.org/Zagolo/danger-kotlin/pull-requests/1"
+        }
+      },
+      "title": "README.md edited online with Bitbucket"
+    },
+    "content": {
+      "raw": "Test Comment",
+      "markup": "markdown",
+      "html": "<p>Test Comment</p>",
+      "type": "rendered"
+    },
+    "created_on": 1603196142,
+    "user": {
+      "username": "f-meloni",
+      "display_name": "Franco Meloni",
+      "type": "user",
+      "uuid": "{bd1991e4-a3ed-45b2-be38-acea659650f1}",
+      "links": {
+        "self": {
+          "href": "https://api.bitbucket.org/2.0/users/%7Bbd1991e4-a3ed-45b2-be38-acea659650f1%7D"
+        },
+        "html": {
+          "href": "https://bitbucket.org/%7Bbd1991e4-a3ed-45b2-be38-acea659650f1%7D/"
+        },
+        "avatar": {
+          "href": "https://secure.gravatar.com/avatar/3d90e967de2beab6d44cfadbb4976b87?d=https%3A%2F%2Favatar-management--avatars.us-west-2.prod.public.atl-paas.net%2Finitials%2FFM-1.png"
+        }
+      }
+    },
+    "updated_on": 1603196142,
+    "type": "pullrequest_comment",
+    "id": 109197958
+  }
+"""